Done with this. Merging to main

+++ FFXIV-RaidLootAPI/Controllers/PlayerController.cs
@@ -106,7 +106,7 @@ namespace FFXIV_RaidLootAPI.Controllers
             return false;
         }
 
-        async private Task<bool> UserIsAuthorized(HttpContext HttpContext, string playerId, DataContext context){
+        async public Task<bool> UserIsAuthorized(HttpContext HttpContext, string playerId, DataContext context){
             //Console.WriteLine("Checking authorization");
             if (HttpContext.Request.Cookies.TryGetValue("jwt_xivloot", out var jwt)){
                 
@@ -751,9 +751,14 @@ namespace FFXIV_RaidLootAPI.Controllers
         {
             using (var context = _context.CreateDbContext())
             {   
+
                 Players? player = await context.Players.FindAsync(id);
                 if (player is null)
                     return NotFound("Player not found");
+
+                bool isAuthorized = await UserIsAuthorized(HttpContext, player.Id.ToString(), context); // Even if unclaimed player must be in static to delete it.
+                if(!isAuthorized)
+                    return Unauthorized("Not Authorized");
 
                 context.Players.Remove(player);
                 await context.SaveChangesAsync();
@@ -768,6 +773,13 @@ namespace FFXIV_RaidLootAPI.Controllers
                 Players? player = await context.Players.FindAsync(id);
                 if (player is null)
                     return NotFound("Player not found");
+
+            if (player.IsClaimed)
+            {   
+                bool isAuthorized = await UserIsAuthorized(HttpContext, player.Id.ToString(), context);
+                if(!isAuthorized)
+                    return Unauthorized("Not Authorized");
+            }
 
                 player.IsAlt = !player.IsAlt;
                 await context.SaveChangesAsync();
